Safari-609.1.11.1/LayoutTests/js/large-expressions-expected.txt

This test checks behavior with expressions that create deep parse trees.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS eval(repeatedExpression("letterA", "+", 100)) is repeatedString("a", 100)
PASS eval(repeatedExpression("letterA", "+", 1000)) is repeatedString("a", 1000)
PASS eval(repeatedExpression("letterA", "+", 100000)) threw exception Error: Out of memory.
PASS successfullyParsed is true

TEST COMPLETE
